+++ resources/schema.json
@@ -207,16 +207,16 @@
                 "IdenticalEqual": { "$ref": "#/definitions/default-mutator-config" },
                 "InstanceOf_": { "$ref": "#/definitions/default-mutator-config" },
                 "LogicalAnd": { "$ref": "#/definitions/default-mutator-config" },
-                "LogicalAndNegateAllSubExpr": { "$ref": "#/definitions/default-mutator-config" },
-                "LogicalAndNegateSingleSubExpr": { "$ref": "#/definitions/default-mutator-config" },
+                "LogicalAndAllSubExprNegation": { "$ref": "#/definitions/default-mutator-config" },
                 "LogicalAndNegation": { "$ref": "#/definitions/default-mutator-config" },
+                "LogicalAndSingleSubExprNegation": { "$ref": "#/definitions/default-mutator-config" },
                 "LogicalLowerAnd": { "$ref": "#/definitions/default-mutator-config" },
                 "LogicalLowerOr": { "$ref": "#/definitions/default-mutator-config" },
                 "LogicalNot": { "$ref": "#/definitions/default-mutator-config" },
                 "LogicalOr": { "$ref": "#/definitions/default-mutator-config" },
-                "LogicalOrNegateAllSubExpr": { "$ref": "#/definitions/default-mutator-config" },
-                "LogicalOrNegateSingleSubExpr": { "$ref": "#/definitions/default-mutator-config" },
+                "LogicalOrAllSubExprNegation": { "$ref": "#/definitions/default-mutator-config" },
                 "LogicalOrNegation": { "$ref": "#/definitions/default-mutator-config" },
+                "LogicalOrSingleSubExprNegation": { "$ref": "#/definitions/default-mutator-config" },
                 "NotEqualNotIdentical": { "$ref": "#/definitions/default-mutator-config" },
                 "NotIdenticalNotEqual": { "$ref": "#/definitions/default-mutator-config" },
                 "TrueValue": {
