+++ src/Mutator/ProfileList.php
@@ -100,16 +100,16 @@ final class ProfileList
         // IdenticalEqual disabled from the default boolean profile
         Mutator\Boolean\InstanceOf_::class,
         Mutator\Boolean\LogicalAnd::class,
-        Mutator\Boolean\LogicalAndNegateAllSubExpr::class,
-        Mutator\Boolean\LogicalAndNegateSingleSubExpr::class,
+        Mutator\Boolean\LogicalAndAllSubExprNegation::class,
         Mutator\Boolean\LogicalAndNegation::class,
+        Mutator\Boolean\LogicalAndSingleSubExprNegation::class,
         Mutator\Boolean\LogicalLowerAnd::class,
         Mutator\Boolean\LogicalLowerOr::class,
         Mutator\Boolean\LogicalNot::class,
         Mutator\Boolean\LogicalOr::class,
-        Mutator\Boolean\LogicalOrNegateAllSubExpr::class,
-        Mutator\Boolean\LogicalOrNegateSingleSubExpr::class,
+        Mutator\Boolean\LogicalOrAllSubExprNegation::class,
         Mutator\Boolean\LogicalOrNegation::class,
+        Mutator\Boolean\LogicalOrSingleSubExprNegation::class,
         // NotEqualNotIdentical disabled from the default boolean profile
         // NotIdenticalNotEqual disabled from the default boolean profile
         Mutator\Boolean\TrueValue::class,
@@ -332,16 +332,16 @@ final class ProfileList
         'IdenticalEqual' => Mutator\Boolean\IdenticalEqual::class,
         'InstanceOf_' => Mutator\Boolean\InstanceOf_::class,
         'LogicalAnd' => Mutator\Boolean\LogicalAnd::class,
-        'LogicalAndNegateAllSubExpr' => Mutator\Boolean\LogicalAndNegateAllSubExpr::class,
-        'LogicalAndNegateSingleSubExpr' => Mutator\Boolean\LogicalAndNegateSingleSubExpr::class,
+        'LogicalAndAllSubExprNegation' => Mutator\Boolean\LogicalAndAllSubExprNegation::class,
         'LogicalAndNegation' => Mutator\Boolean\LogicalAndNegation::class,
+        'LogicalAndSingleSubExprNegation' => Mutator\Boolean\LogicalAndSingleSubExprNegation::class,
         'LogicalLowerAnd' => Mutator\Boolean\LogicalLowerAnd::class,
         'LogicalLowerOr' => Mutator\Boolean\LogicalLowerOr::class,
         'LogicalNot' => Mutator\Boolean\LogicalNot::class,
         'LogicalOr' => Mutator\Boolean\LogicalOr::class,
-        'LogicalOrNegateAllSubExpr' => Mutator\Boolean\LogicalOrNegateAllSubExpr::class,
-        'LogicalOrNegateSingleSubExpr' => Mutator\Boolean\LogicalOrNegateSingleSubExpr::class,
+        'LogicalOrAllSubExprNegation' => Mutator\Boolean\LogicalOrAllSubExprNegation::class,
         'LogicalOrNegation' => Mutator\Boolean\LogicalOrNegation::class,
+        'LogicalOrSingleSubExprNegation' => Mutator\Boolean\LogicalOrSingleSubExprNegation::class,
         'NotEqualNotIdentical' => Mutator\Boolean\NotEqualNotIdentical::class,
         'NotIdenticalNotEqual' => Mutator\Boolean\NotIdenticalNotEqual::class,
         'TrueValue' => Mutator\Boolean\TrueValue::class,
